Make latex notation for best value and confidence boundaries.

Make latex notation for best value and confidence boundaries.

latex_range(best, min, max, ndigits = "auto", na = "-")

Arguments

best

best value

min

lower boundary value

max

upper boundary value

ndigits

'auto' (default) or number of digits after decimal point

na

show NA value as na. default na='-'

Value

character vector

Examples

x <- c(NA,1:10, NA)
x <- x + rnorm(length(x))
err <- rnorm(length(x))

latex_range(best = x, min = x - err, max = x + err, ndigits = 2)
#>  [1] "$-$"             "$0.60\\pm2.03$"  "$0.63\\pm-0.59$" "$3.16\\pm-0.68$"
#>  [5] "$2.66\\pm-0.77$" "$5.27\\pm0.65$"  "$6.55\\pm0.80$"  "$6.63\\pm-0.17$"
#>  [9] "$8.05\\pm-1.07$" "$7.46\\pm0.69$"  "$9.54\\pm-0.77$" "$-$"            

latex_range(best = x, min = x - err, max = x + 2*err, ndigits = 2, na='')
#>  [1] "$$"                       "$0.60_{-2.03}^{+4.05}$"  
#>  [3] "$0.63_{--0.59}^{+-1.17}$" "$3.16_{--0.68}^{+-1.37}$"
#>  [5] "$2.66_{--0.77}^{+-1.54}$" "$5.27_{-0.65}^{+1.30}$"  
#>  [7] "$6.55_{-0.80}^{+1.60}$"   "$6.63_{--0.17}^{+-0.34}$"
#>  [9] "$8.05_{--1.07}^{+-2.14}$" "$7.46_{-0.69}^{+1.37}$"  
#> [11] "$9.54_{--0.77}^{+-1.53}$" "$$"