Make latex notation for best value and confidence boundaries.

Make latex notation for best value and confidence boundaries.

latex_range(best, min, max, ndigits = 0, na = "-")

Arguments

best

best value

min

lower boundary value

max

upper boundary value

ndigits

number of digits after decimal point

na

show NA value as na. default na='-'

Value

character vector

Examples

x <- c(NA,1:10, NA)
x <- x + rnorm(length(x))
err <- rnorm(length(x))

latex_range(best = x, min = x - err, max = x + err, ndigits = 2)
#> Warning: Is argument `ndigits` too small?
#>  [1] "$-$"              "$1.67\\pm0.62$"   "$3.94\\pm-0.00$"  "$3.72\\pm-0.22$" 
#>  [5] "$3.83\\pm-0.39$"  "$5.90\\pm2.44$"   "$4.93\\pm-1.02$"  "$4.58\\pm-0.14$" 
#>  [9] "$7.83\\pm-0.71$"  "$11.01\\pm-0.42$" "$8.86\\pm-0.47$"  "$-$"             

latex_range(best = x, min = x - err, max = x + 2*err, ndigits = 2, na='')
#> Warning: Is argument `ndigits` too small?
#>  [1] "$$"                      "$1.67_{-0.62}^{+1.25}$" 
#>  [3] "$3.94_{0.00}^{+-0.01}$"  "$3.72_{0.22}^{+-0.44}$" 
#>  [5] "$3.83_{0.39}^{+-0.79}$"  "$5.90_{-2.44}^{+4.88}$" 
#>  [7] "$4.93_{1.02}^{+-2.04}$"  "$4.58_{0.14}^{+-0.29}$" 
#>  [9] "$7.83_{0.71}^{+-1.42}$"  "$11.01_{0.42}^{+-0.84}$"
#> [11] "$8.86_{0.47}^{+-0.94}$"  "$$"